backend/requirements.txt
fastapi
httpx
numpy
pandas
pydantic
python-dotenv
redis
uvicorn
geopy